添加传感器数据包的接收和显示，下一步添加蓝牙自动连接上一次连接的设备

+++ user_info.cpp
@@ -6,7 +6,7 @@ user_info::user_info()
 
 }
 
-user_info::user_info(QString nmae, int age, double height, double weight) :
+user_info::user_info(QString nmae, int age, float height, float weight) :
     name(nmae), age(age), height(height), weight(weight)
 {
 }
